// src/shared/types/store.ts
export enum StoreTheme {
  MINIMAL_LIGHT = "MINIMAL_LIGHT",
  MINIMAL_DARK = "MINIMAL_DARK",
  BOLD_LIGHT = "BOLD_LIGHT",
  BOLD_DARK = "BOLD_DARK",
  CLASSIC = "CLASSIC",
}

export type Store = {
  id: string;
  username: string;
  name: string;
  bio: string;
  logoUrl: string;
  bannerUrl: string;
  theme: StoreTheme;
  socialLinks: {
    instagram: string;
    whatsapp: string;
    facebook: string;
    twitter: string;
  };
  createdAt: string;
};